requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.29
djangorestframework==3.9.4
drf-yasg==1.17.1
idna==2.10
importlib-metadata==2.1.3
inflection==0.3.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.9
pathlib2==2.3.7.post1
pyparsing==2.4.7
pytz==2022.1
redis==3.5.3
requests==2.27.1
ruamel.ordereddict==0.4.15
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.9
vine==1.3.0
zipp==1.2.0
